extract test_exec_env to a separate file
